chore: move files to layout folder to not create new ones
refs #1030
